feat: Create normalizeObjectArray.js
Given an array of objects, each with identical property names but different values, combines them into one object with each property name associated with an array of the given objects' values.

Example:

```
const data = [{ a: 1, b: 'a' }, { a: 2, b: 'b' }]
console.log(normalizeObjectArray(data)) // { a: [1, 2], b: ['a', 'b'] }
```

If passed an empty array, this will return an empty array. It's already squashed.